refactor(funput-ffi): modularize and guard the C boundary against panics
Split the 320-LOC lib.rs into per-concern modules, each under 150 LOC:
- config.rs   — the 7 engine setters
- compose.rs  — process_char/backspace/flip_composing/clear/arm/buffer
- shortcuts.rs — add/clear_shortcuts (gõ tắt)
- support.rs  — safe() panic guard
- lib.rs      — module root: opaque handle, new/free, flat re-export surface

Wrap every entry point that touches the engine in support::safe()
(catch_unwind), so a panic in funput-engine/funput-core is caught at the
boundary and returned as a no-op result instead of unwinding into the host.
Under edition 2024 such an unwind aborts the process — for an IME that means
killing the user's keyboard mid-typing. The guard turns an engine bug into a
single dropped keystroke. Mirrors the existing funput-jni::support::safe.

The C ABI is unchanged: the set of exported signatures is identical; only the
declaration order and doc comments in the regenerated funput.h differ.

Also add an English README (README.en.md) with a language nav on both files
for open-source discoverability.

Verified: cargo test -p funput-ffi (5 unit + 12 round_trip) and
cargo clippy -p funput-ffi --all-targets -- -D warnings both green.
