+++ crates/funput-ffi/README.en.md
@@ -1,0 +1,126 @@
+# funput-ffi
+
+**English** · [Tiếng Việt](README.md)
+
+The **C ABI** boundary for `funput-engine` — lets a **non-Rust** shell drive the engine through C
+functions. The engine ships in a `.dylib`/`.so`/`.a`; the native side (Swift, C++) loads and calls it.
+
+> A **Rust** consumer (the Windows shell, `funput-cli`) links `funput-engine` directly and does **not**
+> need this crate. FFI is only for **macOS** (Swift IMKit) and the **Fcitx5 addon on Linux** (C++).
+
+## What this crate does
+
+Only **boundary marshalling**: `extern "C"` + `#[repr(C)]`, converting `ImeResult` (Rust) ↔
+`FunputResult` (C), plus null-safety. **No** Telex/VNI logic, **no** hook/inject — those belong to the
+engine and the platform.
+
+## C API (`include/funput.h`)
+
+Handle-based; results are returned **by value** (POD, no free needed); input is a **codepoint** (the
+platform maps keycode → char itself). Every function is **null-safe** (a null handle / invalid codepoint
+yields a `None` result) and **panic-safe**: every entry point that touches the engine runs inside
+`support::safe()` (`catch_unwind`), so a panic in the engine is caught at the boundary and turned into a
+no-op result — it **never** unwinds into the host (which would abort the whole IME process).
+
+```c
+typedef struct FunputEngine FunputEngine;   // opaque handle
+
+typedef struct {
+    uint8_t  action;        // 0=None, 1=Send, 2=Restore
+    uint32_t backspace;     // chars to delete before inserting
+    uint32_t count;         // valid codepoints in chars (<= 64)
+    uint32_t chars[64];     // UTF-32 output; chars[0..count] are valid
+} FunputResult;
+
+FunputEngine *funput_engine_new(void);
+void          funput_engine_free(FunputEngine *engine);
+
+void          funput_set_method(FunputEngine *engine, uint8_t method);      // 0=Telex, 1=VNI
+void          funput_set_tone_style(FunputEngine *engine, uint8_t style);   // 0=Traditional, 1=Modern
+void          funput_set_enabled(FunputEngine *engine, bool enabled);
+void          funput_set_smart_restore(FunputEngine *engine, bool on);
+void          funput_set_eager_restore(FunputEngine *engine, bool on);
+void          funput_clear(FunputEngine *engine);                            // word boundary / focus change
+
+FunputResult  funput_process_char(FunputEngine *engine, uint32_t codepoint);
+FunputResult  funput_backspace(FunputEngine *engine);                       // Backspace while composing
+uintptr_t     funput_buffer(const FunputEngine *engine, uint32_t *out, uintptr_t cap); // copy composing buffer (UTF-32) into out, returns char count
+```
+
+Applying a result: `action == 0 (None)` → let the app receive the key as usual; otherwise delete
+`backspace` chars, then insert `chars[0..count]`. `funput_buffer` lets the platform render the
+preedit/marked text from the composing buffer.
+
+The header is generated by **cbindgen** (committed). Regenerate after changing the `extern "C"` surface:
+
+```bash
+bash scripts/gen-header.sh    # requires: cargo install cbindgen
+```
+
+## Marshalling (`src/types.rs`)
+
+`FunputResult::from_ime(&ImeResult)`:
+- `Action::{None, Send, Restore}` → `0 / 1 / 2`.
+- `output.chars()` → `chars[..count]`, **truncated** at `CHARS_CAP = 64` (the overflow policy lives
+  here, not in the engine).
+- `backspace as u32`. Input via `char::from_u32(codepoint)`; `None` → empty result.
+
+## Memory ownership
+
+| Side | Responsibility |
+|------|----------------|
+| Rust (`funput_engine_new`) | Allocates the handle |
+| Caller (Swift/C++) | Calls `funput_engine_free()` exactly **once** per handle |
+| `funput_process_char` / `funput_backspace` | Return **by value** — no allocation, no per-result free |
+
+Only the **handle** needs freeing (Swift typically `deinit { funput_engine_free(handle) }`). A result is
+POD on the stack → no leak.
+
+## macOS flow (example)
+
+```
+IMKInputController.handle (Swift)
+   └─ keycode → codepoint
+      funput_process_char(engine, cp)        ← funput-ffi
+         └─ funput-engine → FunputResult (by value)
+            Swift reads action / backspace / chars[0..count]
+            → setMarkedText / insertText      ← outside funput-ffi
+```
+
+## Layout & build
+
+```
+src/lib.rs          # module root: opaque FunputEngine + new/free + re-export surface
+src/config.rs       # 7 setters: method/tone_style/enabled/smart|eager_restore/spell/autocap
+src/compose.rs      # process_char/backspace/flip_composing/clear/arm/buffer
+src/shortcuts.rs    # add_shortcut/clear_shortcuts (text expansion, "gõ tắt")
+src/support.rs      # safe(): catch_unwind guard for the C boundary
+src/types.rs        # #[repr(C)] FunputResult + from_ime() + CHARS_CAP/ACTION_*
+cbindgen.toml
+scripts/gen-header.sh
+include/funput.h     # GENERATED (committed)
+```
+
+`crate-type = ["cdylib", "staticlib", "rlib"]`. Artifacts: macOS `libfunput_ffi.a`/`.dylib` + header
+(built via `platforms/macos/scripts/build-ffi.sh`); Windows does not use it (the shell links the engine
+directly); the Linux Fcitx5 addon links `libfunput_ffi` + includes `funput.h`.
+
+Edition 2024 note: use `#[unsafe(no_mangle)]` and explicit `unsafe { }` around `Box::from_raw` /
+`ptr.as_mut()`.
+
+## Dependencies & callers
+
+- `funput-ffi → funput-engine → funput-core`.
+- Consumers: `platforms/macos` (Swift, bridging header) and `platforms/linux/fcitx5` (C++,
+  `ffi_handle.h`). **Not** used by: `funput-cli`, the Windows shell (both link the engine directly).
+
+## Tests
+
+```bash
+cargo test  -p funput-ffi
+cargo clippy -p funput-ffi --all-targets -- -D warnings
+cargo build -p funput-ffi && ls target/debug/libfunput_ffi.*   # .a .dylib .rlib
+```
+
+`src/types.rs` (unit: `from_ime`, truncate > 64) + `tests/round_trip.rs` (drives the `extern "C"` API
+like a C caller: Telex/VNI/English-restore, null-safety, surrogate).
